# Implement Nop FrameGen Writer