--- a/gsv_test.go
+++ b/gsv_test.go
@@ -103,9 +103,13 @@ func Benchmark_sleep_sort(b *testing.B)     { benchmarkSort("sleep", b) }
 func Benchmark_stooge_sort(b *testing.B)    { benchmarkSort("stooge", b) }
 func Benchmark_quick_sort(b *testing.B)     { benchmarkSort("quick", b) }
 
+// WriteNop is a writer for FrameGen that does nothing.
+// Ensures we only benchmark alogrithms.
+func WriteNop(_ []int) {}
+
 func benchmarkSort(sort string, b *testing.B) {
 	arr := randomArray(Count, Max)
-	frameGen := FrameGen(WriteStdout)
+	frameGen := FrameGen(WriteNop)
 	if sort, found := sorterMap[sort]; found {
 		for n := 0; n < b.N; n++ {
 			sort(arr, frameGen)